moved test cases
