+++ tests/resources/3-links/1-agent-full-leg.binpb
