+++ tests/resources/3-links/3-links-config-2.yml
@@ -1,10 +1,10 @@
 modules:
   protofiles:
     type: ProtoFiles
-    network: ./assets/3-links/3-links-network.binpb
-    population: ./assets/3-links/1-agent-full-leg.binpb
-    vehicles: ./assets/3-links/vehicles.binpb
-    ids: ./assets/3-links/ids.binpb
+    network: ./tests/resources/3-links/3-links-network.binpb
+    population: ./tests/resources/3-links/1-agent-full-leg.binpb
+    vehicles: ./tests/resources/3-links/vehicles.binpb
+    ids: ./tests/resources/3-links/ids.binpb
   partitioning:
     type: Partitioning
     num_parts: 2
