+++ tests/resources/3-links/3-links-network.binpb
