+++ tests/resources/3-links/ids.binpb
