+++ tests/resources/3-links/vehicles.binpb
