+++ tests/resources/adhoc_routing/no_updates/agents.binpb
