+++ tests/resources/adhoc_routing/no_updates/ids.binpb
