+++ tests/resources/adhoc_routing/no_updates/network.binpb
