--- tests/resources/adhoc_routing/no_updates/output_events.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<events version="1.0">
-    <event time="32400" type="actend" person="0" link="0" actType="home_28800.0" />
-    <event time="32400" type="departure" person="0" link="0" legMode="walk" />
-    <event time="32408" type="travelled" person="0" distance="10" mode="walk" />
-    <event time="32409" type="actend" person="0" link="0" actType="car interaction" />
-    <event time="32409" type="departure" person="0" link="0" legMode="car" />
-    <event time="32409" type="PersonEntersVehicle" person="0" vehicle="0" />
-    <event time="32419" type="left link" link="0" vehicle="0" />
-    <event time="32419" type="entered link" link="5" vehicle="0" />
-    <event time="32469" type="left link" link="5" vehicle="0" />
-    <event time="32469" type="entered link" link="6" vehicle="0" />
-    <event time="32519" type="left link" link="6" vehicle="0" />
-    <event time="32519" type="entered link" link="4" vehicle="0" />
-    <event time="32529" type="PersonLeavesVehicle" person="0" vehicle="0" />
-    <event time="32529" type="arrival" person="0" link="4" legMode="car" />
-    <event time="32529" type="actstart" person="0" link="4" actType="car interaction" />
-    <event time="32530" type="actend" person="0" link="4" actType="car interaction" />
-    <event time="32530" type="departure" person="0" link="4" legMode="walk" />
-    <event time="32538" type="travelled" person="0" distance="10" mode="walk" />
-    <event time="33600" type="actend" person="1" link="0" actType="home_28800.0" />
-    <event time="33600" type="departure" person="1" link="0" legMode="walk" />
-    <event time="33608" type="travelled" person="1" distance="10" mode="walk" />
-    <event time="33609" type="actend" person="1" link="0" actType="car interaction" />
-    <event time="33609" type="departure" person="1" link="0" legMode="car" />
-    <event time="33609" type="PersonEntersVehicle" person="1" vehicle="1" />
-    <event time="33619" type="left link" link="0" vehicle="1" />
-    <event time="33619" type="entered link" link="1" vehicle="1" />
-    <event time="33629" type="left link" link="1" vehicle="1" />
-    <event time="33629" type="entered link" link="2" vehicle="1" />
-    <event time="33729" type="left link" link="2" vehicle="1" />
-    <event time="33729" type="entered link" link="3" vehicle="1" />
-    <event time="33739" type="PersonLeavesVehicle" person="1" vehicle="1" />
-    <event time="33739" type="arrival" person="1" link="3" legMode="car" />
-    <event time="33739" type="actstart" person="1" link="3" actType="car interaction" />
-    <event time="33740" type="actend" person="1" link="3" actType="car interaction" />
-    <event time="33740" type="departure" person="1" link="3" legMode="walk" />
-    <event time="33748" type="travelled" person="1" distance="10" mode="walk" />
-    <event time="34200" type="actend" person="2" link="0" actType="home_28800.0" />
-    <event time="34200" type="departure" person="2" link="0" legMode="walk" />
-    <event time="34208" type="travelled" person="2" distance="10" mode="walk" />
-    <event time="34209" type="actend" person="2" link="0" actType="car interaction" />
-    <event time="34209" type="departure" person="2" link="0" legMode="car" />
-    <event time="34209" type="PersonEntersVehicle" person="2" vehicle="2" />
-    <event time="34219" type="left link" link="0" vehicle="2" />
-    <event time="34219" type="entered link" link="5" vehicle="2" />
-    <event time="34269" type="left link" link="5" vehicle="2" />
-    <event time="34269" type="entered link" link="6" vehicle="2" />
-    <event time="34319" type="left link" link="6" vehicle="2" />
-    <event time="34319" type="entered link" link="4" vehicle="2" />
-    <event time="34329" type="PersonLeavesVehicle" person="2" vehicle="2" />
-    <event time="34329" type="arrival" person="2" link="4" legMode="car" />
-    <event time="34329" type="actstart" person="2" link="4" actType="car interaction" />
-    <event time="34330" type="actend" person="2" link="4" actType="car interaction" />
-    <event time="34330" type="departure" person="2" link="4" legMode="walk" />
-    <event time="34338" type="travelled" person="2" distance="10" mode="walk" />
-    <event time="34800" type="actend" person="3" link="5" actType="home_28800.0" />
-    <event time="34800" type="departure" person="3" link="5" legMode="walk" />
-    <event time="34807" type="travelled" person="3" distance="8.944272" mode="walk" />
-    <event time="34808" type="actend" person="3" link="5" actType="car interaction" />
-    <event time="34808" type="departure" person="3" link="5" legMode="car" />
-    <event time="34808" type="PersonEntersVehicle" person="3" vehicle="3" />
-    <event time="34858" type="left link" link="5" vehicle="3" />
-    <event time="34858" type="entered link" link="6" vehicle="3" />
-    <event time="34908" type="left link" link="6" vehicle="3" />
-    <event time="34908" type="entered link" link="4" vehicle="3" />
-    <event time="34918" type="PersonLeavesVehicle" person="3" vehicle="3" />
-    <event time="34918" type="arrival" person="3" link="4" legMode="car" />
-    <event time="34918" type="actstart" person="3" link="4" actType="car interaction" />
-    <event time="34919" type="actend" person="3" link="4" actType="car interaction" />
-    <event time="34919" type="departure" person="3" link="4" legMode="walk" />
-    <event time="34927" type="travelled" person="3" distance="10" mode="walk" />
-</events>
