+++ tests/resources/adhoc_routing/no_updates/vehicles.binpb
