+++ tests/resources/adhoc_routing/with_updates/agents.binpb
