+++ tests/resources/adhoc_routing/with_updates/config-1.yml
@@ -1,10 +1,10 @@
 modules:
   protofiles:
     type: ProtoFiles
-    network: ./assets/adhoc_routing/with_updates/network.binpb
-    population: ./assets/adhoc_routing/with_updates/agents.binpb
-    vehicles: ./assets/adhoc_routing/with_updates/vehicles.binpb
-    ids: ./assets/adhoc_routing/with_updates/ids.binpb
+    network: ./tests/resources/adhoc_routing/with_updates/network.binpb
+    population: ./tests/resources/adhoc_routing/with_updates/agents.binpb
+    vehicles: ./tests/resources/adhoc_routing/with_updates/vehicles.binpb
+    ids: ./tests/resources/adhoc_routing/with_updates/ids.binpb
   partitioning:
     type: Partitioning
     num_parts: 1
