+++ tests/resources/adhoc_routing/with_updates/ids.binpb
