+++ tests/resources/adhoc_routing/with_updates/network.binpb
