+++ tests/resources/adhoc_routing/with_updates/vehicles.binpb
