--- tests/resources/three_link/full_legs/output_events.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<events version="1.0">
-    <event time="32400" type="actend" person="0" link="0" actType="home_28800.0" />
-    <event time="32400" type="departure" person="0" link="0" legMode="walk" />
-    <event time="32408" type="travelled" person="0" distance="10" mode="walk" />
-    <event time="32409" type="actend" person="0" link="0" actType="car interaction" />
-    <event time="32409" type="departure" person="0" link="0" legMode="car" />
-    <event time="32409" type="PersonEntersVehicle" person="0" vehicle="0" />
-    <event time="32419" type="left link" link="0" vehicle="0" />
-    <event time="32419" type="entered link" link="1" vehicle="0" />
-    <event time="32519" type="left link" link="1" vehicle="0" />
-    <event time="32519" type="entered link" link="2" vehicle="0" />
-    <event time="32529" type="PersonLeavesVehicle" person="0" vehicle="0" />
-    <event time="32529" type="arrival" person="0" link="2" legMode="car" />
-    <event time="32529" type="actstart" person="0" link="2" actType="car interaction" />
-    <event time="32530" type="actend" person="0" link="2" actType="car interaction" />
-    <event time="32530" type="departure" person="0" link="2" legMode="walk" />
-    <event time="32546" type="travelled" person="0" distance="20" mode="walk" />
-</events>
