--- tests/resources/three_link/multiple_agents/no_jam/output_events.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<events version="1.0">
-    <event time="32400" type="actend" person="0" link="0" actType="home_28800.0" />
-    <event time="32400" type="departure" person="0" link="0" legMode="car" />
-    <event time="32400" type="PersonEntersVehicle" person="0" vehicle="0" />
-    <event time="32410" type="left link" link="0" vehicle="0" />
-    <event time="32410" type="entered link" link="1" vehicle="0" />
-    <event time="32510" type="left link" link="1" vehicle="0" />
-    <event time="32510" type="entered link" link="2" vehicle="0" />
-    <event time="32520" type="PersonLeavesVehicle" person="0" vehicle="0" />
-    <event time="32520" type="arrival" person="0" link="2" legMode="car" />
-    <event time="32520" type="actstart" person="0" link="2" actType="errands_2400.0" />
-    <event time="33000" type="actend" person="1" link="0" actType="home_28800.0" />
-    <event time="33000" type="departure" person="1" link="0" legMode="bike" />
-    <event time="33000" type="PersonEntersVehicle" person="1" vehicle="1" />
-    <event time="33020" type="left link" link="0" vehicle="1" />
-    <event time="33020" type="entered link" link="1" vehicle="1" />
-    <event time="33220" type="left link" link="1" vehicle="1" />
-    <event time="33220" type="entered link" link="2" vehicle="1" />
-    <event time="33240" type="PersonLeavesVehicle" person="1" vehicle="1" />
-    <event time="33240" type="arrival" person="1" link="2" legMode="bike" />
-    <event time="33240" type="actstart" person="1" link="2" actType="errands_2400.0" />
-    <event time="33600" type="actend" person="2" link="0" actType="home_28800.0" />
-    <event time="33600" type="departure" person="2" link="0" legMode="car" />
-    <event time="33600" type="PersonEntersVehicle" person="2" vehicle="2" />
-    <event time="33610" type="left link" link="0" vehicle="2" />
-    <event time="33610" type="entered link" link="1" vehicle="2" />
-    <event time="33710" type="left link" link="1" vehicle="2" />
-    <event time="33710" type="entered link" link="2" vehicle="2" />
-    <event time="33720" type="PersonLeavesVehicle" person="2" vehicle="2" />
-    <event time="33720" type="arrival" person="2" link="2" legMode="car" />
-    <event time="33720" type="actstart" person="2" link="2" actType="errands_2400.0" />
-</events>
