--- tests/resources/three_link/output_events.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<events version="1.0">
-<event time="32400" type="actend" person="0" link="0" actType="home_28800.0" />
-<event time="32400" type="departure" person="0" link="0" legMode="car" />
-<event time="32400" type="PersonEntersVehicle" person="0" vehicle="0" />
-<event time="32410" type="left link" link="0" vehicle="0" />
-<event time="32410" type="entered link" link="1" vehicle="0" />
-<event time="32510" type="left link" link="1" vehicle="0" />
-<event time="32510" type="entered link" link="2" vehicle="0" />
-<event time="32520" type="PersonLeavesVehicle" person="0" vehicle="0" />
-<event time="32520" type="arrival" person="0" link="2" legMode="car" />
-<event time="32520" type="actstart" person="0" link="2" actType="errands_2400.0" />
-</events>
